+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - Dependency graph inefficiency that caused excessive compilation time
 - Removed JSON stream readers which are much slower than strings and vectors
+- Filtered out several EVM-specific warnings from standard JSON output
 - File path resolution for EraVM-specific warnings in standard JSON output
 
 ## [1.5.0] - 2024-06-10